[FIX] dod: Rework userstyle to match updated Odoo docs
Odoo Documentation was already updated a while ago but the old
userstyle didn't match with this version of the docs.

With this commit, the userstyle is updated to match the new
version of the docs.

In the meantime, as the old documentation is deprecated (not
supported by Odoo itself), the old version of the userstyle that
matched the old documentation is now unmaintaned and
dropped completely.